+++ src/peigs/ctof/pdspgv_f.c
@@ -157,7 +157,7 @@ void pdspgv_(ifact, n, matrixA, mapA, matrixB, mapB, matZ, mapZ, eval,
        *info = linfo;
      
      fprintf( stderr, " %s me = %d argument %d is a pointer to NULL. \n",
-              msg, me, -linfo );
+              msg, (int)me, (int)(-linfo) );
      l_exit_( &linfo, msg );
      return;
    }
@@ -169,7 +169,7 @@ void pdspgv_(ifact, n, matrixA, mapA, matrixB, mapB, matZ, mapZ, eval,
     if ( *ibuffsz < (nvecsA + nvecsB + nvecsZ + 3 )) {
 	*info = -10;
 	
-        fprintf( stderr, " %s me = %d ibuffsz is too small. \n", msg, me );
+        fprintf( stderr, " %s me = %d ibuffsz is too small. \n", msg, (int)me );
 	l_exit_( info, msg );
 	return;
       }
