+++ src/peigs/src/c/b_ortho.c
@@ -188,7 +188,7 @@ void b_ortho ( n, colB, mapB, m, colZ, mapZ, ibuffptr, iwork, work, ort, info)
   if( *n < *m ) {
       fprintf(stderr,
               "Error in routine b_ortho. m (=%d) < n (=%d). me = %d \n",
-              *m, *n, me);
+              (int)(*m), (int)(*n), (int)me);
       exit(-1);
   }
